+++ DataStorage/Storage - Display.cpp
@@ -8,7 +8,7 @@ std::vector<Data> DisplayStorage::displayList;
 //for display command
 std::vector<Data> & DisplayStorage::getDisplayList(TimeMacro startTime, TimeMacro endTime){
 	displayList.clear(); //clear the vector everytime it starts
-	std::vector<int> timePeriod;
+	std::vector<long long> timePeriod;
 
 	timePeriod = DataBase::searchPeriod(startTime,endTime);
 	int endT;
